requirements.txt
# ============================================================
# Requirements para MBA IA - Desafio Ingestão e Busca
# ============================================================
# Projeto: Sistema RAG com LangChain, Gemini e pgVector
# Versão: 1.0.0
# Python: 3.10+
# ============================================================

# Core - LangChain Framework
langchain==0.3.27
langchain-core==0.3.74
langchain-text-splitters==0.3.9

# LangChain - Integrações Específicas
langchain-google-genai==2.1.9        # Google Gemini (Embeddings + LLM)
langchain-postgres==0.0.15           # PostgreSQL + pgVector
langchain-community==0.3.27          # Document loaders (PyPDFLoader)

# Database - PostgreSQL
psycopg[binary]==3.2.9               # PostgreSQL driver moderno
asyncpg==0.30.0                      # Async support para langchain-postgres

# Environment & Configuration
python-dotenv==1.1.1                 # Gerenciamento de variáveis de ambiente

# PDF Processing
pypdf==6.0.0                         # Dependência do PyPDFLoader

# Google AI
google-ai-generativelanguage==0.6.18 # Gemini
google-auth==2.40.0                  # Autenticação Google